Move Example class from examples.string_ops to examples package
